Upgrade numpy (#754)
* Bump numpy from 1.26.4 to 2.1.0

Bumps [numpy](https://github.com/numpy/numpy) from 1.26.4 to 2.1.0.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/main/doc/RELEASE_WALKTHROUGH.rst)
- [Commits](https://github.com/numpy/numpy/compare/v1.26.4...v2.1.0)

---
updated-dependencies:
- dependency-name: numpy
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* tutorials

* docstring failures

* correct some failures

* disable cvxpy related failures

* temporary disable rst checks

* try downgrading to 2.0.0

* add back line to test tutorials

* uncomment 2.1.0 failin examples

* update lock file

---------

+++ toqito/state_metrics/fidelity.py
@@ -50,7 +50,7 @@ def fidelity(rho: np.ndarray, sigma: np.ndarray) -> float:
     ... )
     >>> sigma = rho
     >>> fidelity(rho, sigma)
-    1.0000000000000002
+    np.float64(1.0000000000000002)
 
     References
     ==========
